refactor: move ProcessorEngine from ports/shared_lib to application/engine/
ProcessorEngine is an application-layer orchestrator, not a port — it owns
the filter pipeline, YOLO detector cache and all filter dispatch logic.
Moving it to application/engine/ makes the layer ownership explicit.

- git mv processor_engine.{h,cpp} → application/engine/
- Change namespace jrb::ports::shared_lib → jrb::application::engine
- Create application/engine/BUILD (cc_library :engine)
- Delete ports/shared_lib/ (entirely removed — cgo era is over)
- Update all callers: processor_engine_adapter.{h,cpp}, webrtc_manager.{h,cpp},
  main_client.cpp — include paths and namespace references updated throughout
- Update ports/grpc/BUILD deps to use //application/engine:engine
