+++ src/cpp_accelerator/application/engine/processor_engine.h
@@ -14,7 +14,7 @@
 #include "src/cpp_accelerator/domain/interfaces/grayscale_algorithm.h"
 #include "src/cpp_accelerator/infrastructure/cuda/i_yolo_detector.h"
 
-namespace jrb::ports::shared_lib {
+namespace jrb::application::engine {
 
 class ProcessorEngine {
 public:
@@ -42,4 +42,4 @@ private:
       detector_cache_;
 };
 
-}  // namespace jrb::ports::shared_lib
+}  // namespace jrb::application::engine
